moved `as` dyn casters to their own utility header
