+++ src/ast/Attribute.cpp
@@ -7,8 +7,7 @@
  */
 
 #include "ast/Attribute.h"
-#include "souffle/utility/MiscUtil.h"
-
+#include "souffle/utility/DynamicCasting.h"
 #include <ostream>
 #include <utility>
