+++ src/ast/Constant.cpp
@@ -7,7 +7,7 @@
  */
 
 #include "ast/Constant.h"
-#include "souffle/utility/MiscUtil.h"
+#include "souffle/utility/DynamicCasting.h"
 #include <ostream>
 #include <utility>
