+++ src/ast/Pragma.cpp
@@ -7,7 +7,7 @@
  */
 
 #include "ast/Pragma.h"
-#include "souffle/utility/MiscUtil.h"
+#include "souffle/utility/DynamicCasting.h"
 #include <ostream>
 #include <utility>
