+++ src/ast/Variable.cpp
@@ -7,7 +7,7 @@
  */
 
 #include "ast/Variable.h"
-#include "souffle/utility/MiscUtil.h"
+#include "souffle/utility/DynamicCasting.h"
 #include <ostream>
 #include <utility>
