+++ src/include/souffle/utility/FunctionalUtil.h
@@ -16,6 +16,7 @@
 
 #pragma once
 
+#include "souffle/utility/DynamicCasting.h"
 #include <algorithm>
 #include <functional>
 #include <set>
