+++ src/include/souffle/utility/MiscUtil.h
@@ -186,75 +186,6 @@ bool equal_ptr(const Own<T>& a, const Own<T>& b) {
     return equal_ptr(a.get(), b.get());
 }
 
-/**
- * This class is used to tell as<> that cross-casting is allowed.
- * I use a named type rather than just a bool to make the code stand out.
- */
-class AllowCrossCast {};
-
-/**
- * Helpers for `dynamic_cast`ing without having to specify redundant type qualifiers.
- * e.g. `as<AstLiteral>(p)` instead of `as<AstLiteral>(p)`.
- */
-template <typename B, typename CastType = void, typename A>
-auto as(A* x) {
-    if constexpr (!std::is_same_v<CastType, AllowCrossCast>) {
-        static_assert(std::is_base_of_v<std::remove_const_t<A>, std::remove_const_t<B>>,
-                "`as<B, A>` does not allow cross-type dyn casts. "
-                "(i.e. `as<B, A>` where `B <: A` is not true.) "
-                "Such a cast is likely a mistake or typo.");
-    }
-    return dynamic_cast<copy_const_t<A, B>*>(x);
-}
-
-template <typename B, typename CastType = void, typename A>
-auto as(A& x) {
-    return as<B, CastType>(&x);
-}
-
-template <typename B, typename CastType = void, typename A>
-auto as(Own<A>& x) {
-    return as<B, CastType>(x.get());
-}
-
-template <typename B, typename CastType = void, typename A>
-auto as(const Own<A>& x) {
-    return as<B, CastType>(x.get());
-}
-
-/**
- * Down-casts and checks the cast has succeeded
- */
-template <typename B, typename CastType = void, typename A>
-auto& asAssert(A&& a) {
-    auto* cast = as<B, CastType>(std::forward<A>(a));
-    assert(cast && "Invalid cast");
-    return *cast;
-}
-
-/**
- * Checks if the object of type Source can be casted to type Destination.
- */
-template <typename B, typename A>
-bool isA(A* x) {
-    // Dont't forward onto as<> - need to check cross-casting
-    return dynamic_cast<copy_const_t<A, B>*>(x) != nullptr;
-}
-
-template <typename B, typename A>
-auto isA(A& x) {
-    return isA<B>(&x);
-}
-
-template <typename B, typename A>
-bool isA(const Own<A>& x) {
-    return isA<B>(x.get());
-}
-
-template <typename B, typename A>
-bool isA(Own<A>& x) {
-    return isA<B>(x.get());
-}
 // -------------------------------------------------------------------------------
 //                               Error Utilities
 // -------------------------------------------------------------------------------
